refactor getSmsManager to be a standalone function
