+++ android/build.gradle
@@ -2,5 +2,5 @@
 plugins {
     id("com.android.application") version "7.4.2" apply false
     id("com.android.library") version "7.4.2" apply false
-    id("org.jetbrains.kotlin.android") version "1.8.0" apply false
+    id("org.jetbrains.kotlin.android") version '1.9.20-Beta' apply false
 }
